requirements.txt
# InkTrace 后端依赖
# Python 3.8+ 环境要求

## 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-core==2.14.0

## 数据处理与解析
pandas==2.1.4
numpy==1.26.2
python-multipart==0.0.6

## AI与机器学习
openai==1.3.9
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4

## 文本处理
jieba==0.42.1
langchain==0.0.349
langchain-community==0.0.10

## 网络与HTTP
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

## 工具与工具包
tqdm==4.66.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

## 开发与测试
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

## 系统工具
psutil==5.9.6

# 可选依赖（根据需求安装）
# tensorflow==2.15.0  # 如果需要TensorFlow支持
# scikit-learn==1.3.2  # 如果需要机器学习功能
# matplotlib==3.8.2  # 如果需要可视化

# 环境特定依赖
# 生产环境可能需要：
# gunicorn==21.2.0
# supervisor==4.2.5